# Simplify detection of x86 CPU features