… build_tools/fbcode_config.sh
… build_tools/fbcode_config_platform010.sh