@socketpair 2023-05-02
Anyway. I still don't understand why we can just add runtime detection? Why not to:
1. have several `.c` files implementing the same algo, but using different implementations.
2. `-march xxx` should cut some of them off. I mean, if a given arch guarantee some instruction set, the slower implementations should be omitted.
3. each `.c` file from the list above, should be compiled with different compiler options.
4. The main function (when run first time) should detect the best implementation and fill the global function pointer to one of the implementations.

In other words, setting `-march xxx` should change (influence to) the list of available implementations.

Why this is important ? For common distributives, like Fedora, Debian or Ubuntu. Unfortunately, they have to support relatively old hardware. Everything may look like:

(I imply the compiler supports all high-end intrinsics, so we don't have to check it at build stage)

low-end, middle-end and high-end - just pseudocode that helps understand my vision.


|                     |   Low-end impl   |  Middle-end impl | High-end impl    |
|---------------------|:----------------:|-----------------:|------------------|
| PORTABLE=low-end    | Assume supported | Check at runtime | Check at runtime |
| PORTABLE=middle-end |       Omit       | Assume supported | Check at runtime |
| PORTABLE=high-end   |       Omit       |             Omit | Assume supported |


some-end-impl.c (example for one of the algo implementations. Should be compiled with appropriate `-march` without respect of PORTABLE=XXX):
```C
int some_end_impl(int x) {
#if PORTABLE > SOME_END_ARCH  // pseudocode.
    assert(0);
#else
    (implementation for SOME_END ARCH)
#endif
}

int some_end_supported() { 
#if PORTABLE > SOME_END_ARCH
    return 0; // useless to include since more fast implmentation is guaranteed to be supported by choosing PORTABLE=XXX
#elif PORTABLE == SOME_END_ARCH
    return 1; // don't need to check since PORTABLE=XXX guarantee this is supported.
#else
    // this may or may not be supported by CPU at runtime. check it
    (check at runtime)
#endif
}
```

general_code.c:
```C
int algo(int x) {
    static int (*impl)(int x) = NULL;
    if (!impl) {
        if (high_end_supported())
            impl=high_end_impl;
        else if (middle_end_supported())
            impl=middle_end_impl;
        else if (low_end_supported())
            impl=low_end_impl;
        else
            assert(0);
    }
    return impl(x)
}
```