@mrambacher 2023-05-02
> > Anyway. I still don't understand why we can just add runtime detection? Why not to:
> > ...
> > Why this is important ? For common distributives, like Fedora, Debian or Ubuntu. Unfortunately, they have to support relatively old hardware.
> 
> If crc32c were the default checksum, you would have a point. It's not the default. And I think you're dug into the out-of-date premise that this is a singularly performance-critical component. For most of our workloads, block checksums are less than 1% of RocksDB CPU. Essentially all are less than 5%. Doubling that cost in a portable build is not going to be the only overhead associated with a portable build. Lots of other things passively benefit from the compiler using newer, wider instructions.
> 
> And anyway we have never had runtime detection of AVX2 and switching between xxh3 implementations. So this change is not a regression in default behaviors under any build config.
> 
> Under this change, everyone (using default checksum) still gets reasonable performance in a portable build. If you are sensitive to RocksDB performance, you should build specifically for your hardware, and that relates to more than just checksum functions.

My concern is what happens in binary distributions (such as Java).  Can those binary distributions still take advantage of the architectural efficiencies?  Personally, I would prefer we did more "stuff" at run-time (AVX and other support, URING detection, run-time detection of compression libraries, etc) that would allow these features to be used and discovered in binary distributions.