@pdillinger 2023-05-03
I believe a single java jar could can contain various binary builds to be chosen at runtime. With kXXH3 as default checksum, the most significant x86 CPU feature is probably AVX2. If you want to support pre-AVX2 hardware, you could probably just bundle two binaries, one for "haswell" and newer/greater, and one for the compiler minimum. Though you might need a small native library itself to identify which RocksDB library is compatible. Runtime checking at that level (select which full library) should be much more reliable and maintainable than the stuff I'm trying to get rid of, even if there is extra weight required in compilation time and package size.

Although it's tempting to try to do clever stuff like only compile certain translation units with various supported instruction sets, so you don't have to include two full RocksDB libraries, I believe it's possible to run into ODR or linker issues when mixing different -march settings. Probably not in the current code (unless you're using Folly F14, in which case you will!) but someday you might, and better safe than sorry.