@pdillinger 2023-05-03
> While I agree you could package multiple SO for Java, you are then pushing a run-time determination on which to use into Java, which I do not know why would be any easier/better than making that same determination in a single library in C++.

Like I said, you could use a small bootstrap native library to make the determination.

> For the machine intrinsics, my understanding is that you can call most (if not all) of those intrinsics even if the architecture does not support them. Doing so will result in an Illegal Instruction exception and not a link error. If there are ODR issues, then the methods should be refactored anyway for testing purposes.

But then you're limited to the benefits of the hand-coded optimizations implemented with runtime selection, not those opportunities found by your compiler nor compile-time-selected hand-coded optimizations in dependences such as std or folly.

> I fully support simplifying the rules here but feel that compile-time choices (instead of run-time) can lead to a bunch of other issues and makes it hard to develop and deploy features that may be available on some hardware or installations.

Can you be more specific? As evidenced by recurring bug fix PRs, runtime detection is a testing nightmare and involves a level of compile time detection also. I don't see your point.